Now handles floating point pitches

+++ src/pv-normalize.py
@@ -8,7 +8,7 @@ for file in sys.argv[1:]:
     filename, fileext = os.path.splitext(file)
     if fileext == '.pv':
         print('Processing ' + filename + '.pv', end = '\r')
-        pv = np.array([int(line.strip()) for line in open(file, 'r')])
+        pv = np.array([float(line.strip()) for line in open(file, 'r')])
         avg = np.mean(pv)
         std = np.std(pv)
         npv = (pv - avg)/std
